+++ src/lib/pdf/LogbookPdf.tsx
@@ -79,7 +79,7 @@ const s = StyleSheet.create({
 function formatDate(iso: string): string {
   const [y, m, d] = iso.split('-').map(Number);
   if (!y || !m || !d) return iso || '—';
-  return new Date(y, m - 1, d).toLocaleDateString('en-GB', { day: '2-digit', month: 'short', year: 'numeric' });
+  return new Date(y, m - 1, d).toLocaleDateString('en-GB', { day: '2-digit', month: 'long', year: 'numeric' });
 }
 
 function RankTable({ title, items }: { title: string; items: RankItem[] }) {
